backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database and ODM
motor==3.3.2
beanie==1.23.6
pymongo==4.6.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-decouple==3.8

# Validation and Serialization
pydantic==2.5.0
pydantic-settings==2.0.3

# Export functionality
python-docx==1.1.0
reportlab==4.0.7
markdown==3.5.1

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
isort==5.12.0

# CORS and middleware  
# CORS is built into FastAPI, no separate package needed
